+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- Rename "inputs" to "deps"
 
 ## [1.3.0] - 2019-03-04
 ### Changed